can roll 2 cubes for table games
