+++ Week2/app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">Dice Roller</string>
-    <string name="roll">Roll</string>
+    <string name="roll1">Roll_one_cube</string>
+    <string name="roll2">Roll_two_cubes</string>
 </resources>
